+++ src/Task.h
@@ -2,6 +2,7 @@
 #ifndef TASK_H
 #define TASK_H 
 #include <string>
+#include <chrono>
 
 enum Status { Pending, InProgress, Completed };
 
@@ -12,6 +13,8 @@ class Task
         std::string title;
         std::string description;
         Status status;
+        std::chrono::system_clock::time_point createdAt;
+        std::chrono::system_clock::time_point deadline;
         
     public:
         Task(int id, const std::string& title, const std::string& desc);
@@ -27,9 +30,18 @@ class Task
         std::string getTitle() const;
         std::string getDescription() const;
         Status getStatus() const;
+        std::chrono::system_clock::time_point getCreatedAt() const;
+        std::chrono::system_clock::time_point getDeadline() const;
+        
         void setTitle(const std::string& newTitle);
         void setDescription(const std::string& newDesc);
         void setStatus(Status newStatus);
+        void setDeadline(std::chrono::system_clock::time_point dl);
+        void setCreatedAt(std::chrono::system_clock::time_point ca);
+        
+        bool isOverdue() const;
+        std::string getCreatedAtStr() const;
+        std::string getDeadlineStr() const;
 }; 
 
 #endif // TASK_H